+++ app/Repositories/QuoteFile/QuoteFileRepository.php
@@ -7,18 +7,19 @@ use App\Contracts\Repositories\QuoteFile\{
     ImportableColumnRepositoryInterface as ImportableColumnRepository,
     FileFormatRepositoryInterface as FileFormatRepository
 };
+use App\Enum\Lock;
 use App\Models\{
-    Quote\BaseQuote as Quote,
     QuoteFile\QuoteFile,
     QuoteFile\DataSelectSeparator,
     QuoteFile\ImportedRow
 };
 use App\Services\PdfParser\PdfOptions;
-use Illuminate\Http\UploadedFile;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Arr;
-use Illuminate\Support\Collection as SupportCollection;
-use Storage, Str, File, DB;
+use Illuminate\Contracts\Cache\LockTimeoutException;
+use Illuminate\Database\Query\Builder as BaseBuilder;
+use Illuminate\Database\Eloquent\Collection as DbCollection;
+use Illuminate\Support\{Arr, Str, Collection, Facades\Storage, Facades\File, Facades\DB, Facades\Cache};
+use Throwable;
+use Webpatser\Uuid\Uuid;
 
 class QuoteFileRepository implements QuoteFileRepositoryInterface
 {
@@ -30,7 +31,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
     protected ImportableColumnRepository $importableColumn;
 
-    protected static ?Collection $importableColumnsCache = null;
+    protected static ?DbCollection $importableColumnsCache = null;
 
     public function __construct(
         QuoteFile $quoteFile,
@@ -83,33 +84,6 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $quoteFile->load('dataSelectSeparator')->makeHidden('user');
     }
 
-    public function createPdf(Quote $quote, array $attributes)
-    {
-        if (!isset($attributes['original_file_path']) || !isset($attributes['filename'])) {
-            return null;
-        }
-
-        $original_file_path = $attributes['original_file_path'];
-        $quote_file = new UploadedFile(Storage::path($attributes['original_file_path']), $attributes['filename']);
-        $format = $this->fileFormat->whereInExtension(['pdf']);
-        $quote_id = $quote->id;
-        $file_type = 'Generated PDF';
-
-        $quote->generatedPdf()->delete();
-
-        $attributes = compact('quote_file', 'format', 'file_type', 'original_file_path', 'quote_id');
-
-        if (app()->runningInConsole()) {
-            $attributes = array_merge($attributes, ['user' => $quote->user]);
-        }
-
-        $quoteFile = $this->create($attributes);
-
-        $quote->load('generatedPdf');
-
-        return $quoteFile;
-    }
-
     public function createRawData(QuoteFile $quoteFile, array $array)
     {
         $user = $quoteFile->user;
@@ -140,35 +114,37 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $quoteFile->importedRawData()->orderBy('page')->get();
     }
 
-    public function createRowsData(QuoteFile $quoteFile, array $array)
-    {
-        /**
-         * Delete early imported data
-         */
-        DB::transaction(
-            fn () => $quoteFile->rowsData()->forceDelete(),
-            5
-        );
-
-        $this->createImportedPages($array, $quoteFile);
-    }
-
     public function createScheduleData(QuoteFile $quoteFile, array $value)
     {
         $user = $quoteFile->user;
 
-        /**
-         * Delete early imported payment schedule data
-         */
-        $quoteFile->scheduleData()->forceDelete();
+        $lock = Cache::lock(Lock::UPDATE_QUOTE_FILE($quoteFile->getKey()), 10);
 
-        $scheduleData = $quoteFile->scheduleData()->make(compact('value'));
-        $scheduleData->user()->associate($user);
-        $scheduleData->save();
+        $lock->block(30);
 
-        $quoteFile->markAsHandled();
+        DB::beginTransaction();
 
-        return $scheduleData;
+        try {
+            if ($quoteFile->scheduleData()->exists()) {
+                $quoteFile->scheduleData()->forceDelete();
+            }
+
+            $scheduleData = $quoteFile->scheduleData()->make(compact('value'));
+            $scheduleData->user()->associate($user);
+            $scheduleData->save();
+
+            $quoteFile->markAsHandled();
+
+            DB::commit();
+
+            return $scheduleData;
+        } catch (Throwable $e) {
+            DB::rollBack();
+
+            throw $e;
+        } finally {
+            optional($lock)->release();
+        }
     }
 
     public function getRowsData(QuoteFile $quoteFile)
@@ -183,7 +159,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
 
     public function find(string $id)
     {
-        return $this->quoteFile->query()->whereId($id)->firstOrFail();
+        return $this->quoteFile->query()->whereKey($id)->firstOrFail();
     }
 
     public function findByClause(array $clause)
@@ -191,44 +167,28 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         return $this->quoteFile->query()->where($clause)->first();
     }
 
+    public function findByQuote(string $quoteID, string $fileType)
+    {
+        return $this->quoteFile->query()
+            ->whereExists(fn (BaseBuilder $builder) =>
+            $builder->selectRaw('1')->from('quotes')
+                ->where('id', $quoteID)
+                ->where(fn (BaseBuilder $builder) => $builder->whereColumn('quote_files.id', 'quotes.distributor_file_id')->orWhereColumn('quote_files.id', 'quotes.schedule_file_id')))
+            ->where('file_type', $fileType)
+            ->first();
+    }
+
     public function exists(string $id)
     {
         return $this->quoteFile->whereId($id)->exists();
     }
 
-    public function deletePriceListsExcept(QuoteFile $quoteFile)
-    {
-        $exceptedQuoteFileId = $quoteFile->id;
-        $quote = $quoteFile->quote;
-
-        return $quote->quoteFiles()->priceLists()->whereKeyNot($exceptedQuoteFileId)->delete();
-    }
-
-    public function deletePaymentSchedulesExcept(QuoteFile $quoteFile)
-    {
-        $exceptedQuoteFileId = $quoteFile->id;
-        $quote = $quoteFile->quote;
-
-        return $quote->quoteFiles()->paymentSchedules()->whereKeyNot($exceptedQuoteFileId)->delete();
-    }
-
-    public function deleteExcept(QuoteFile $quoteFile)
-    {
-        if ($quoteFile->isSchedule()) {
-            return $this->deletePaymentSchedulesExcept($quoteFile);
-        }
-
-        return $this->deletePriceListsExcept($quoteFile);
-    }
-
     public function replicatePriceList(QuoteFile $quoteFile, ?string $quoteId = null): QuoteFile
     {
         $quote_file_id = $quoteFile->id;
-        $user_id = $quoteFile->user_id;
 
         /** @var QuoteFile */
         $quoteFileCopy = tap($quoteFile->replicate(), function ($file) use ($quoteId) {
-            $file->quote()->associate($quoteId);
             $file->save();
         });
 
@@ -238,9 +198,9 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         $importedRowColumn = 'replicated_row_id';
 
         /** Generating new Ids for Imported Columns and Rows in the temporary table. */
-        DB::select(
+        DB::statement(
             "create temporary table `{$tempRowsTable}`
-            select uuid() as `id`, `id` as {$importedRowColumn}, `quote_file_id`, `columns_data`, `page`, `is_selected`, `group_name` from `imported_rows`
+            select uuid() as `id`, `id` as {$importedRowColumn}, `quote_file_id`, `columns_data`, `page`, `is_selected` from `imported_rows`
                 where `imported_rows`.`quote_file_id` = :quote_file_id",
             compact('quote_file_id')
         );
@@ -248,15 +208,12 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         /** We are marking temporary rows as selected if the related payload is present in the current request. */
         $this->updateTempSelectedImportedRows($tempRowsTable, $importedRowColumn);
 
-        /** We are updating temporary group description if the related payload is present in the current request. */
-        // $this->updateTempGroupDescription($tempRowsTable, $importedRowColumn);
-
         /** Inserting Imported Rows with new Ids. */
         DB::insert(
-            "insert into `imported_rows` (`id`, `{$importedRowColumn}`, `user_id`, `quote_file_id`, `columns_data`, `page`, `is_selected`, `group_name`)
-            select `id`, `{$importedRowColumn}`, :user_id, :new_quote_file_id, `columns_data`, `page`, `is_selected`, `group_name`
+            "insert into `imported_rows` (`id`, `{$importedRowColumn}`,`quote_file_id`, `columns_data`, `page`, `is_selected`)
+            select `id`, `{$importedRowColumn}`, :new_quote_file_id, `columns_data`, `page`, `is_selected`
             from `{$tempRowsTable}`",
-            compact('user_id', 'new_quote_file_id')
+            compact('new_quote_file_id')
         );
 
         return $quoteFileCopy;
@@ -295,16 +252,20 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
             array_push($importedRows, ...$this->createImportedRows($rows, $quoteFile, $page));
         });
 
-        if (count($importedRows) < 1) {
-            $quoteFile->setException(QFNRF_01, 'QFNRF_01');
-            $quoteFile->markAsUnHandled();
-            $quoteFile->throwExceptionIfExists();
+        if (empty($importedRows)) {
+            return;
         }
 
-        $quoteFile->rowsData()->saveMany($importedRows);
+        $rows = array_map(fn (ImportedRow $row) => $row->getAttributes() + [
+            'id' => (string) Uuid::generate(4),
+            'created_at' => now()->toDateTimeString(),
+            'updated_at' => now()->toDateTimeString(),
+        ], $importedRows);
+
+        ImportedRow::insert($rows);
     }
 
-    protected function createImportedRows(array $rows, QuoteFile $quoteFile, int $page): SupportCollection
+    protected function createImportedRows(array $rows, QuoteFile $quoteFile, int $page): Collection
     {
         $quote_file_id = $quoteFile->id;
         $user_id = $quoteFile->user->id;
@@ -319,13 +280,13 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
             ));
     }
 
-    protected function mapColumnsData(array $row): SupportCollection
+    protected function mapColumnsData(array $row): Collection
     {
         $importableColumns = $this->getImportableColumns();
 
         Arr::forget($row, PdfOptions::SYSTEM_HEADER_ONE_PAY);
 
-        return SupportCollection::wrap($row)->map(
+        return Collection::wrap($row)->map(
             fn ($value, $name) => [
                 'header' => Str::header($name),
                 'value' => $value,
@@ -334,7 +295,7 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         )->values();
     }
 
-    protected function getImportableColumns(): Collection
+    protected function getImportableColumns(): DbCollection
     {
         if (isset(static::$importableColumnsCache)) {
             return static::$importableColumnsCache;
@@ -357,32 +318,5 @@ class QuoteFileRepository implements QuoteFileRepositoryInterface
         DB::table($table)->update(['is_selected' => false]);
 
         DB::table($table)->{$updatableScope}($column, $selectedRowsIds)->update(['is_selected' => true]);
-    }
-
-    private function updateTempGroupDescription(string $table, string $column): void
-    {
-        if (!request()->is('api/quotes/groups/*')) {
-            return;
-        }
-
-        $rowsIds = request()->input('rows', []);
-        $name = request()->input('name');
-
-        /** Create Group Description request */
-        if (request()->isMethod('post')) {
-            DB::table($table)->whereIn($column, $rowsIds)->update(['group_name' => $name]);
-        }
-
-        /** Update Group Description request */
-        if (request()->isMethod('patch')) {
-            DB::table($table)->where('group_name', request()->group_name)->update(['group_name' => null]);
-            DB::table($table)->whereIn($column, $rowsIds)->update(['group_name' => $name]);
-        }
-
-        /** Move Group Description rows request */
-        if (request()->isMethod('put')) {
-            DB::table($table)->where('group_name', request()->from_group_name)->whereIn($column, $rowsIds)
-                ->update(['group_name' => request()->to_group_name]);
-        }
     }
 }
